+++ src/api/office/index.ts
@@ -1,2 +1,3 @@
+export { FontFaceDeclarations } from './FontFaceDeclarations';
 export { TextBody } from './TextBody';
 export { TextDocument } from './TextDocument';